requirements.txt
aniso8601==7.0.0
asgiref==3.5.0
dj-database-url==0.5.0
Django==3.2
django-filter==21.1
django-graphql-jwt==0.3.0
django-heroku==0.3.1
djangorestframework==3.13.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
iso8601==1.0.2
promise==2.3
psycopg2==2.9.3
PyJWT==1.7.1
pytz==2021.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
whitenoise==6.0.0
